Sprint 3 temp

+++ src/main/java/com/ust/pos/model/Price.java
@@ -4,11 +4,13 @@ import jakarta.persistence.Entity;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @Setter
 public class Price extends CommonFields {
     private String product;
-    private Double priceAmount;
+    private BigDecimal priceAmount;
     private String priceType;
 }
